Add files via upload
Temple ruins had to be maped because of how directions it goes where most other elvel go up or left to right, Why there are lines is because my scanner cuts out some of the words near the edge if there isn't enough on the paper there. this is the research for Upstream, Hog Wilde, The Lost City, Temple Ruins & Road to Nowhere.